[refactor] Remove Hive and Hadoop dependencies from SQL Client
