+++ flink-end-to-end-tests/flink-end-to-end-tests-hive/src/test/resources/META-INF/services/org.apache.flink.table.factories.Factory 
@@ -13,6 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-org.apache.flink.table.client.gateway.utils.SimpleCatalogFactory
-org.apache.flink.table.client.gateway.local.DependencyTest$TestCatalogFactory
-org.apache.flink.table.client.gateway.context.ExecutionContextTest$TestClassLoaderCatalogFactory
+org.apache.flink.tests.util.hive.TestHiveCatalogFactory
