+++ flink-end-to-end-tests/flink-end-to-end-tests-hive/src/test/resources/log4j2-test.properties
@@ -1,0 +1,1 @@
+
