+++ flink-end-to-end-tests/flink-end-to-end-tests-hive/src/test/resources/test-sql-client-catalogs.yaml
